fix(message-signing): add working utility fn
