+++ packages/common/src/index.ts
@@ -4,4 +4,5 @@ export * from './fetchUtil';
 export * from './logger';
 export * from './utils';
 export * from './constants';
+export * from './signatures';
 export * from './keys';
